Residual Network (ResNet 1d)

We implement 1d Residual Networks.

Note

ResNet is an example of a sequence-to-point, non-recurrent architecture. See foreword for more details.

IMUDevNNArchitectures.ResNet1d.resnet — Function

resnet(in_out::Pair{Tuple{Int,Int},Int}, group_sizes;
       strides=vcat([1], fill(2, length(group_sizes) - 1)),
       base_plane::Int=64, kernel_size::Int=3,
       outputblock_builder=(in_out) -> outputblock(in_out;
                                                   hidden=1024,
                                                   dropout=0.5,
                                                   transition_size=0))

ResNet1d model (see the original ResNet publication). It consists of an input block, a residual block, and an output block. The residual block is defined by the group_sizes, strides, base_plane and kernel_size. The output block is often custom built, but a default one is provided. in_out argument in the outputblock_builder is a pair of the form (num_channels, time_dim) => output_dim.

Arguments

in_out::Pair{Tuple{Int,Int},Int}: The input and output sizes of the model.

group_sizes: The number of basic blocks in each residual group.

strides: The stride for the first basic block in each residual group.

base_plane::Int: Hyperparameter based on which the number of channels in each residual group is going to be computed.

kernel_size::Int: The size of the kernel in each basic block.

outputblock_builder: A function that builds the output block of the model.

Note

The in_out has a format (num_channels, time_dim) => output_dim.

IMUDevNNArchitectures.ResNet1d.outputblock — Function

outputblock(in_out::Pair{Tuple{Int,Int},Int};
            hidden::Int=1024,
            dropout::Float64=0.5,
            transition_size::Int=0)

Output block for ResNet1d. It can have a transition layer between the residual block and the output layer.

Arguments

in_out::Pair{Tuple{Int,Int},Int}: The input and output sizes of the block.

hidden::Int: The size of the fully connected hidden layers in the output layer.

dropout::Float64: The dropout rate for the output layer.

transition_size::Int: The size of the transition layer.

Note

The in_out has a format (num_channels, time_dim) => output_dim.

IMUDevNNArchitectures.ResNet1d.transition_layer — Function

transition_layer(in_out::Pair{Int,Int})

Transition layer for ResNet1d that can be stuck between the residual block and an output layer. It consists of a convolutional layer and a batch normalization layer.

IMUDevNNArchitectures.ResNet1d.output_layer — Function

output_layer(in_out::Pair{Int,Int};
             hidden::Int=1024,
             dropout::Float64=0.5)

Output layer for ResNet1d. hidden is the size of the fully connected hidden layers.

IMUDevNNArchitectures.ResNet1d.init! — Function

init!(m::Chain)

Initialize the weights and biases of the resnet model m.

Note

If you'd like to zero initialize the weights of the last BatchNormalization in each residual branch of the model, use the function zero_init! after calling this function.

IMUDevNNArchitectures.ResNet1d.zero_init! — Function

Zero-initialize the last BatchNormalization in each residual branch of the model m. This improves the model by 0.2~0.3% according to https://arxiv.org/abs/1706.02677.

Warning

Current implementation is based on the assumption that all SkipConnections in the model m have the same form as the one in basicblock and thus have the BatchNormalization layer as the fourth layer. If this is not the case, the function will not work as expected.

IMUDevNNArchitectures.ResNet1d.inputblock — Function

inputblock(in_out::Pair{Int,Int})

Input block for ResNet1d. It consists of a convolutional layer, a batch normalization layer, and a max pooling layer.

IMUDevNNArchitectures.ResNet1d.residualblock — Function

residualblock(baseplane::Int, groupsizes, strides; kernel_size::Int=3)

Defines the residual block for ResNet1d. It consists of residual groups stacked together.

Arguments

base_plane::Int: Hyperparameter based on which the number of channels in each residual group is going to be computed.

group_sizes: The number of basic blocks in each residual group.

strides: The stride for the first basic block in each residual group.

kernel_size::Int: The size of the kernel in each basic block.

IMUDevNNArchitectures.ResNet1d.residualgroup — Function

residualgroup(kernel_size::Int, in_out::Pair{Int,Int};
              group_size::Int, stride::Int=1)

Residual group for ResNet1d. It consists of basic blocks stacked together. Only the first basic block in the group can have a stride greater than 1.

Arguments

kernel_size::Int: The size of the kernel in each basic block.

in_out::Pair{Int,Int}: The input and output channels of the group.

group_size::Int: The number of basic blocks in the group.

stride::Int: The stride for the first basic block in the group.

IMUDevNNArchitectures.ResNet1d.basicblock — Function

basicblock(kernel_size::Int, in_out::Pair{Int,Int}; stride::Int)

A basic block for ResNet1d. It consists of two convolutional layers with batch normalization and a skip connection. ResNet1d will stack these basic blocks to form residual groups (see also residualgroup).

Arguments

kernel_size::Int: The size of the kernel for the convolutional layers.

in_out::Pair{Int,Int}: The input and output channels of the block.

stride::Int: The stride for the first convolutional layer.